% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Return standard deviation of an input vector/list.}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{vector/list of numeric}
}
\value{
numeric
}
\description{
Return standard deviation of an input vector/list.
}
\examples{
standard_deviation(c(1,2,3,4))
}

