% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Return standard error of an input vector/list.}
\usage{
standard_error(x)
}
\arguments{
\item{x}{vector/list of numeric}
}
\value{
numeric
}
\description{
Return standard error of an input vector/list.
}
\examples{
standard_error(c(1,2,3,4))
}

