% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDeviationR-package.r
\docType{package}
\name{StandardDeviationR}
\alias{StandardDeviationR}
\alias{StandardDeviationR-package}
\title{StandardDeviationR.}
\description{
Calculates the standard deviation of a vector/list input. I made this to practice making packages in R.
}

